+++ src/arch/alpha/utility.hh
@@ -41,7 +41,7 @@
 
 namespace AlphaISA {
 
-uint64_t getArgument(ThreadContext *tc, int &number, uint8_t size,  bool fp);
+uint64_t getArgument(ThreadContext *tc, int &number, uint16_t size, bool fp);
 
 inline bool
 inUserMode(ThreadContext *tc)
