+++ src/arch/sparc/utility.hh
@@ -41,7 +41,8 @@
 
 namespace SparcISA
 {
-    uint64_t getArgument(ThreadContext *tc, int &number, uint8_t size, bool fp);
+    uint64_t
+    getArgument(ThreadContext *tc, int &number, uint16_t size, bool fp);
 
     static inline bool
     inUserMode(ThreadContext *tc)
