+++ src/arch/x86/utility.hh
@@ -52,7 +52,8 @@ class ThreadContext;
 
 namespace X86ISA
 {
-    uint64_t getArgument(ThreadContext *tc, int &number, uint8_t size, bool fp);
+    uint64_t
+    getArgument(ThreadContext *tc, int &number, uint16_t size, bool fp);
 
     static inline bool
     inUserMode(ThreadContext *tc)
